+++ Bnan.Ui/Areas/MAS/Controllers/CarBrandController.cs
@@ -139,7 +139,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
             }
             await SetPageTitleAsync(Status.Insert, pageNumber);
             // Check If code > 9 get error , because code is char(1)
-            if (int.Parse(await GenerateLicenseCodeAsync()) > 3999)
+            if (Int64.Parse(await GenerateLicenseCodeAsync()) > 3999)
             {
                 _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_AddMore"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                 return RedirectToAction("Index", "CarBrand");
@@ -176,7 +176,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
                     return View("AddCarBrand", carBrandVM);
                 }
                 // Check If code > 9 get error , because code is char(1)
-                if (int.Parse(await GenerateLicenseCodeAsync()) > 3999)
+                if (Int64.Parse(await GenerateLicenseCodeAsync()) > 3999)
                 {
                     _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_AddMore"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                     return View("AddCarBrand", carBrandVM);
