+++ Bnan.Ui/Areas/MAS/Controllers/RenterDrivingLicenseController.cs
@@ -120,7 +120,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
                 return RedirectToAction("Index", "RenterDrivingLicense");
             }
             // Check If code > 9 get error , because code is char(1)
-            if (int.Parse(await GenerateLicenseCodeAsync()) > 9)
+            if (Int64.Parse(await GenerateLicenseCodeAsync()) > 9)
             {
                 _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_AddMore"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                 return RedirectToAction("Index", "RenterDrivingLicense");
@@ -159,7 +159,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
                     return View("AddRenterDrivingLicense", renterDrivingLicenseVM);
                 }
                 // Check If code > 9 get error , because code is char(1)
-                if (int.Parse(await GenerateLicenseCodeAsync()) > 9)
+                if (Int64.Parse(await GenerateLicenseCodeAsync()) > 9)
                 {
                     _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_AddMore"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                     return View("AddRenterDrivingLicense", renterDrivingLicenseVM);
@@ -187,7 +187,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
 
             await SetPageTitleAsync(Status.Update, pageNumber);
             // if value with code less than 2 Deleted
-            if (int.Parse(id) < 2 + 1)
+            if (Int64.Parse(id) < 2 + 1)
             {
                 _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_NoUpdate"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                 return RedirectToAction("Index", "RenterDrivingLicense");
@@ -320,12 +320,12 @@ namespace Bnan.Ui.Areas.MAS.Controllers
                     errors.Add(new ErrorResponse { Field = "CrMasSupRenterDrivingLicenseEnName", Message = _localizer["Existing"] });
                 }
                 // Check for existing rental system number
-                else if (existName == "CrMasSupRenterDrivingLicenseNaqlCode" && int.TryParse(dataField, out var code) && code != 0 && All_RenterDrivingLicenses.Any(x => x.CrMasSupRenterDrivingLicenseNaqlCode == code))
+                else if (existName == "CrMasSupRenterDrivingLicenseNaqlCode" && Int64.TryParse(dataField, out var code) && code != 0 && All_RenterDrivingLicenses.Any(x => x.CrMasSupRenterDrivingLicenseNaqlCode == code))
                 {
                     errors.Add(new ErrorResponse { Field = "CrMasSupRenterDrivingLicenseNaqlCode", Message = _localizer["Existing"] });
                 }
                 // Check for existing rental system ID
-                else if (existName == "CrMasSupRenterDrivingLicenseNaqlId" && int.TryParse(dataField, out var id) && id != 0 && All_RenterDrivingLicenses.Any(x => x.CrMasSupRenterDrivingLicenseNaqlId == id))
+                else if (existName == "CrMasSupRenterDrivingLicenseNaqlId" && Int64.TryParse(dataField, out var id) && id != 0 && All_RenterDrivingLicenses.Any(x => x.CrMasSupRenterDrivingLicenseNaqlId == id))
                 {
                     errors.Add(new ErrorResponse { Field = "CrMasSupRenterDrivingLicenseNaqlId", Message = _localizer["Existing"] });
                 }
