+++ Bnan.Ui/Areas/MAS/Controllers/RenterEmployerController.cs
@@ -123,7 +123,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
             }
             await SetPageTitleAsync(Status.Insert, pageNumber);
             // Check If code > 9 get error , because code is char(1)
-            if (int.Parse(await GenerateLicenseCodeAsync()) > 1899999999)
+            if (Int64.Parse(await GenerateLicenseCodeAsync()) > 1899999999)
             {
                 _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_AddMore"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                 return RedirectToAction("Index", "RenterEmployer");
@@ -163,7 +163,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
                     return View("AddRenterEmployer", renterEmployerVM);
                 }
                 // Check If code > 9 get error , because code is char(1)
-                if (int.Parse(await GenerateLicenseCodeAsync()) > 1899999999)
+                if (Int64.Parse(await GenerateLicenseCodeAsync()) > 1899999999)
                 {
                     _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_AddMore"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                     return View("AddRenterEmployer", renterEmployerVM);
@@ -192,7 +192,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
 
             await SetPageTitleAsync(Status.Update, pageNumber);
             // if value with code less than 2 Deleted
-            if (int.Parse(id) < 1800000002 + 1)
+            if (Int64.Parse(id) < 1800000002 + 1)
             {
                 _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_NoUpdate"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                 return RedirectToAction("Index", "RenterEmployer");
