+++ Bnan.Ui/Areas/MAS/Controllers/RenterProfessionController.cs
@@ -123,7 +123,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
             }
             await SetPageTitleAsync(Status.Insert, pageNumber);
             // Check If code > 9 get error , because code is char(1)
-            if (int.Parse(await GenerateLicenseCodeAsync()) > 1499999999)
+            if (Int64.Parse(await GenerateLicenseCodeAsync()) > 1499999999)
             {
                 _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_AddMore"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                 return RedirectToAction("Index", "RenterProfession");
@@ -160,7 +160,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
                     return View("AddRenterProfession", renterProfessionVM);
                 }
                 // Check If code > 9 get error , because code is char(1)
-                if (int.Parse(await GenerateLicenseCodeAsync()) > 1499999999)
+                if (Int64.Parse(await GenerateLicenseCodeAsync()) > 1499999999)
                 {
                     _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_AddMore"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                     return View("AddRenterProfession", renterProfessionVM);
@@ -189,7 +189,7 @@ namespace Bnan.Ui.Areas.MAS.Controllers
 
             await SetPageTitleAsync(Status.Update, pageNumber);
             // if value with code less than 2 Deleted
-            if (int.Parse(id) < 1400000002 + 1)
+            if (Int64.Parse(id) < 1400000002 + 1)
             {
                 _toastNotification.AddErrorToastMessage(_localizer["AuthEmplpoyee_NoUpdate"], new ToastrOptions { PositionClass = _localizer["toastPostion"], Title = "", }); //  إلغاء العنوان الجزء العلوي
                 return RedirectToAction("Index", "RenterProfession");
